Align the code style with project requirements
