+++ src/Cake.Frosting/TaskChains/Chain.cs
@@ -8,20 +8,31 @@ namespace Cake.Frosting.TaskChains
     public static class Chain
     {
         /// <summary>
-        /// Adds the task that will be executed first in the build chain.
+        /// Adds the task by its type that will be executed first in the build chain.
         /// </summary>
         /// <typeparam name="T">The task type.</typeparam>
-        /// <returns>The task in the chain.</returns>
+        /// <returns>Task configuration.</returns>
         public static TaskChainItem Task<T>() where T : IFrostingTask
         {
             return new TaskChainItem { TaskType = typeof(T) };
         }
 
+        /// <summary>
+        /// Adds the task by its name that will be executed first in the build chain.
+        /// </summary>
+        /// <param name="taskName">Name of the task.</param>
+        /// <returns>Task configuration.</returns>
         public static TaskChainItem Task(string taskName)
         {
             return new TaskChainItem { TaskName = taskName };
         }
 
+        /// <summary>
+        /// Adds the task that will be executed after the specified task.
+        /// </summary>
+        /// <typeparam name="T">The type of the task to execute.</typeparam>
+        /// <param name="parent">The parent task after which the task will be executed.</param>
+        /// <returns>Task configuration.</returns>
         public static TaskChainItem Task<T>(this TaskChainItem parent) where T : IFrostingTask
         {
             var nextItem = new TaskChainItem { TaskType = typeof(T) };
@@ -29,6 +40,12 @@ namespace Cake.Frosting.TaskChains
             return nextItem;
         }
 
+        /// <summary>
+        /// Adds the task that will be executed after the specified task.
+        /// </summary>
+        /// <param name="parent">The parent task after which the task will be executed.</param>
+        /// <param name="taskName">Name of the task to execute.</param>
+        /// <returns>Task configuration.</returns>
         public static TaskChainItem Task(this TaskChainItem parent, string taskName)
         {
             var nextItem = new TaskChainItem { TaskName = taskName };
